# src/main/resources/openApi/openApi.yaml
openapi: 3.0.3
info:
  title: Standard API
  version: 1.0.0
servers:
  - url: https://localhost:8080/private/api/v1
...
paths:
  /private/api/v1/brand:
    post:
      summary: Save a new brand
      description: save and returns a new brand if successful.
      tags:
        - BrandController
      operationId: saveBrand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Brand'
      responses:
        '200':
          description: Successful save a new brand
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Brand'
        '400':
          description: Bad request (e.g., missing or invalid fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (e.g., role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: find all brands
      description: return all the brands
      operationId: findAllBrands
      tags:
        - BrandController
      responses:
        '200':
          description: return all the brands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Brand'
        '400':
          description: Bad request (e.g., invalid role)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (e.g., invalid role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /private/api/v1/brand/{id}:
    get:
      summary: find by id
      description: return brand if is exist
      operationId: findBrandById
      tags:
        - BrandController
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: abc123
          description: The ID Brand
      responses:
        '200':
          description: return a brands if exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Brand'
        '400':
          description: Bad request (e.g., invalid role)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (e.g., invalid role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: delete brand by id
      description: return status 200
      operationId: deleteBrandById
      tags:
        - BrandController
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: abc123
          description: The ID Brand
      responses:
        '200':
          description: delete has been successfully executed
        '400':
          description: Bad request (e.g., invalid role)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (e.g., invalid role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: return brand updated brand
      description: return brand updated brand
      operationId: updateBrand
      tags:
        - BrandController
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: abc123
          description: The ID Brand
      requestBody:
        description: Brand to be update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Brand'
      responses:
        '200':
          description: update has been successfully executed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Brand'
        '400':
          description: Bad request (e.g., invalid role)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (e.g., invalid role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /private/api/v1/category:
    post:
      summary: Save a new category
      description: save and returns a new category if successful.
      tags:
        - CategoryController
      operationId: saveCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        '200':
          description: Successful save a new Category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Bad request (e.g., missing or invalid fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (e.g., role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: find all categories
      description: return all the categories
      operationId: findAllCategories
      tags:
        - CategoryController
      responses:
        '200':
          description: return all the categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        '400':
          description: Bad request (e.g., invalid role)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (e.g., invalid role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /private/api/v1/category/{id}:
    get:
      summary: find by id
      description: return category if is exist
      operationId: findCategoryById
      tags:
        - CategoryController
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: abc123
          description: The ID category
      responses:
        '200':
          description: return a categories if exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Bad request (e.g., invalid role)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (e.g., invalid role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: delete brand by id
      description: return status 200
      operationId: deleteCategoryById
      tags:
        - CategoryController
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: abc123
          description: The ID Category
      responses:
        '200':
          description: delete has been successfully executed
        '400':
          description: Bad request (e.g., invalid role)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (e.g., invalid role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: return updated Category
      description: return updated Category
      operationId: updateCategory
      tags:
        - CategoryController
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: abc123
          description: The ID Category
      requestBody:
        description: Category to be update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        '200':
          description: update has been successfully executed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Bad request (e.g., invalid role)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (e.g., invalid role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /private/api/v1/domain:
    post:
      summary: Save a new domain
      description: save and returns a new domain if successful.
      tags:
        - DomainController
      operationId: saveDomain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Domain'
      responses:
        '200':
          description: Successful save a new domain
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Domain'
        '400':
          description: Bad request (e.g., missing or invalid fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (e.g., role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: find all brands
      description: return all the brands
      operationId: findAllDomains
      tags:
        - DomainController
      responses:
        '200':
          description: return all the brands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Domain'
        '400':
          description: Bad request (e.g., invalid role)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (e.g., invalid role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /private/api/v1/domain/{id}:
    get:
      summary: find by id
      description: return domain if is exist
      operationId: findDomainById
      tags:
        - DomainController
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: abc123
          description: The ID Domain
      responses:
        '200':
          description: return a brands if exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Domain'
        '400':
          description: Bad request (e.g., invalid role)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (e.g., invalid role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: delete domain by id
      description: return status 200
      operationId: deleteDomainById
      tags:
        - DomainController
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: abc123
          description: The ID Domain
      responses:
        '200':
          description: delete has been successfully executed
        '400':
          description: Bad request (e.g., invalid role)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (e.g., invalid role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: return updated domain
      description: return updated domain
      operationId: updateDomain
      tags:
        - DomainController
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: abc123
          description: The ID Domain
      requestBody:
        description: Domain to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Domain'
      responses:
        '200':
          description: update has been successfully executed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Domain'
        '400':
          description: Bad request (e.g., invalid role)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (e.g., invalid role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Domain:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        checked:
          type: boolean
        data:
          type: string
          format: date
        hora:
          type: string
          format: date-time
        status:
          type: string
...
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        status:
          type: string
        subcategories:
          type: array
          items:
            $ref: "#/components/schemas/Subcategory"
        date:
          type: string
          format: date
        time:
          type: string
          format: time
    Subcategory:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
            type: string
        status:
            type: string
    Brand:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        status:
          type: string
...
    Error:
      type: object
      properties:
        target:
          type: string
        message:
          type: string
        reference:
          type: string
        timestamp:
          type: string
          format: date-time
        details:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetails"
    ErrorDetails:
      type: object
      properties:
        target:
          type: string
        code:
          type: string
        message:
          type: string